レビュー指摘対応
