+++ src/test/java/nablarch/core/beans/BeanUtilConversionCustomizedTest.java
@@ -22,7 +22,6 @@ import org.junit.runners.Suite;
 @Suite.SuiteClasses({
         BeanUtilConversionCustomizedTest.CopyTest.class,
         BeanUtilConversionCustomizedTest.CreateAndCopyTest.class,
-        BeanUtilConversionCustomizedTest.CreateAndCopyRecordTest.class
 })
 public class BeanUtilConversionCustomizedTest {
 
@@ -576,178 +575,7 @@ public class BeanUtilConversionCustomizedTest {
         }
     }
 
-    /**
-     * {@link BeanUtil#createAndCopy(Class, Object, CopyOptions)}のテスト。
-     * レコードを生成するパターン。
-     * 変換パターンは{@link BeanUtilConversionCustomizedTest.CreateAndCopyTest}でテスト済み。
-     */
-    @SuppressWarnings("NonAsciiCharacters")
-    public static class CreateAndCopyRecordTest {
-
-        @Test
-        public void デフォルト日付パターン() {
-            Src srcBean = new Src();
-            srcBean.setFoo("20180214");
-            srcBean.setBar("20180215");
-            srcBean.setBaz("20180216");
-            CopyOptions copyOptions = CopyOptions.empty();
-            DateDestRecord destRecord = BeanUtil.createAndCopy(DateDestRecord.class, srcBean, copyOptions);
-
-            assertThat(destRecord.foo(), is(date("2018-02-14 00:00:00")));
-            assertThat(destRecord.bar(), is(java.sql.Date.valueOf("2018-02-15")));
-            assertThat(destRecord.baz(), is(Timestamp.valueOf("2018-02-16 00:00:00")));
-        }
-
-        @Test
-        public void デフォルト日付パターンの変換失敗() {
-            Src srcBean = new Src();
-            srcBean.setFoo("2018/02/14");
-            srcBean.setBar("2018/02/15");
-            srcBean.setBaz("2018/02/16");
-            CopyOptions copyOptions = CopyOptions.empty();
-            DateDestRecord destRecord = BeanUtil.createAndCopy(DateDestRecord.class, srcBean, copyOptions);
-
-            assertThat(destRecord.foo(), is(nullValue()));
-            assertThat(destRecord.bar(), is(nullValue()));
-            assertThat(destRecord.baz(), is(nullValue()));
-        }
-
-        @Test
-        public void プロパティ名を指定したカスタム日付パターン() {
-            Src srcBean = new Src();
-            srcBean.setFoo("2018/02/14");
-            srcBean.setBar("2018-02-15");
-            srcBean.setBaz("2018.02.16 12:34:56");
-            CopyOptions copyOptions = CopyOptions.options()
-                    .datePatternByName("foo", "yyyy/MM/dd")
-                    .datePatternByName("bar", "yyyy-MM-dd")
-                    .datePatternByName("baz", "yyyy.MM.dd HH:mm:ss")
-                    .build();
-            DateDestRecord destRecord = BeanUtil.createAndCopy(DateDestRecord.class, srcBean, copyOptions);
-
-            assertThat(destRecord.foo(), is(date("2018-02-14 00:00:00")));
-            assertThat(destRecord.bar(), is(java.sql.Date.valueOf("2018-02-15")));
-            assertThat(destRecord.baz(), is(Timestamp.valueOf("2018-02-16 12:34:56")));
-        }
-
-        @Test
-        public void プロパティ名を指定したカスタムコンバーター() {
-            Src srcBean = new Src();
-            srcBean.setFoo("2018/02/14");
-            srcBean.setBar("2018/02/15");
-            srcBean.setBaz("2018/02/16");
-            final java.util.Date date = new java.util.Date();
-            final java.sql.Date sqlDate = new java.sql.Date(0);
-            final Timestamp timestamp = new Timestamp(0);
-            CopyOptions copyOptions = CopyOptions.options()
-                    .converterByName("foo", java.util.Date.class, value -> date)
-                    .converterByName("bar", java.sql.Date.class, value -> sqlDate)
-                    .converterByName("baz", Timestamp.class, value -> timestamp)
-                    .build();
-            DateDestRecord destRecord = BeanUtil.createAndCopy(DateDestRecord.class, srcBean, copyOptions);
-
-            assertThat(destRecord.foo(), is(sameInstance(date)));
-            assertThat(destRecord.bar(), is(sameInstance(sqlDate)));
-            assertThat(destRecord.baz(), is(sameInstance(timestamp)));
-        }
-
-        @Test
-        public void グローバルなカスタム日付パターン() {
-            Src srcBean = new Src();
-            srcBean.setFoo("2018/02/14");
-            srcBean.setBar("2018/02/15");
-            srcBean.setBaz("2018/02/16");
-            CopyOptions copyOptions = CopyOptions.options()
-                    .datePattern("yyyy/MM/dd").build();
-            DateDestRecord destRecord = BeanUtil.createAndCopy(DateDestRecord.class, srcBean, copyOptions);
-
-            assertThat(destRecord.foo(), is(date("2018-02-14 00:00:00")));
-            assertThat(destRecord.bar(), is(java.sql.Date.valueOf("2018-02-15")));
-            assertThat(destRecord.baz(), is(Timestamp.valueOf("2018-02-16 00:00:00")));
-        }
-
-        @Test
-        public void クラスを指定したカスタムコンバーター() {
-            Src srcBean = new Src();
-            srcBean.setFoo("2018/02/14");
-            srcBean.setBar("2018/02/15");
-            srcBean.setBaz("2018/02/16");
-            final java.util.Date date = new java.util.Date();
-            final java.sql.Date sqlDate = new java.sql.Date(0);
-            final Timestamp timestamp = new Timestamp(0);
-            CopyOptions copyOptions = CopyOptions.options()
-                    .converter(java.util.Date.class, value -> date)
-                    .converterByName("bar", java.sql.Date.class, value -> sqlDate)
-                    .converterByName("baz", Timestamp.class, value -> timestamp)
-                    .build();
-            DateDestRecord destRecord = BeanUtil.createAndCopy(DateDestRecord.class, srcBean, copyOptions);
-
-            assertThat(destRecord.foo(), is(sameInstance(date)));
-            assertThat(destRecord.bar(), is(sameInstance(sqlDate)));
-            assertThat(destRecord.baz(), is(sameInstance(timestamp)));
-        }
-
-        @Test
-        public void デフォルト数字パターン() {
-            Src srcBean = new Src();
-            srcBean.setFoo("1234567890");
-            srcBean.setBar("1234567890");
-            srcBean.setBaz("1234567890");
-            CopyOptions copyOptions = CopyOptions.empty();
-            NumberDestRecord destRecord = BeanUtil.createAndCopy(NumberDestRecord.class, srcBean, copyOptions);
-
-            assertThat(destRecord.foo(), is(1234567890));
-            assertThat(destRecord.bar(), is(1234567890L));
-            assertThat(destRecord.baz(), is(new BigDecimal("1234567890")));
-        }
-
-        @Test
-        public void デフォルト数字パターンの変換失敗() {
-            Src srcBean = new Src();
-            srcBean.setFoo("1,234,567,890");
-            srcBean.setBar("1,234,567,890");
-            srcBean.setBaz("1,234,567,890");
-            CopyOptions copyOptions = CopyOptions.empty();
-            NumberDestRecord destRecord = BeanUtil.createAndCopy(NumberDestRecord.class, srcBean, copyOptions);
-
-            assertThat(destRecord.foo(), is(nullValue()));
-            assertThat(destRecord.bar(), is(nullValue()));
-            assertThat(destRecord.baz(), is(nullValue()));
-        }
-
-        @Test
-        public void プロパティ名を指定したカスタム数字パターン() {
-            Src srcBean = new Src();
-            srcBean.setFoo("1,234,567,890");
-            srcBean.setBar("1,234,567,890");
-            srcBean.setBaz("1,234,567,890");
-            CopyOptions copyOptions = CopyOptions.options()
-                    .numberPatternByName("foo", "#,###")
-                    .build();
-            NumberDestRecord destRecord = BeanUtil.createAndCopy(NumberDestRecord.class, srcBean, copyOptions);
-
-            assertThat(destRecord.foo(), is(1234567890));
-            assertThat(destRecord.bar(), is(nullValue()));
-            assertThat(destRecord.baz(), is(nullValue()));
-        }
-
-        @Test
-        public void グローバルなカスタム数字パターン() {
-            Src srcBean = new Src();
-            srcBean.setFoo("1,234,567,890");
-            srcBean.setBar("1,234,567,890");
-            srcBean.setBaz("1,234,567,890");
-            CopyOptions copyOptions = CopyOptions.options()
-                    .numberPattern("#,###").build();
-            NumberDestRecord destRecord = BeanUtil.createAndCopy(NumberDestRecord.class, srcBean, copyOptions);
-
-            assertThat(destRecord.foo(), is(1234567890));
-            assertThat(destRecord.bar(), is(1234567890L));
-            assertThat(destRecord.baz(), is(new BigDecimal("1234567890")));
-        }
-    }
-
-    private static Date date(String sqlTimestampPattern) {
+    static Date date(String sqlTimestampPattern) {
         return new Date(Timestamp.valueOf(sqlTimestampPattern).getTime());
     }
 
@@ -843,12 +671,4 @@ public class BeanUtilConversionCustomizedTest {
             this.baz = baz;
         }
     }
-
-    public record DateDestRecord(java.util.Date foo,
-                                 java.sql.Date bar,
-                                 Timestamp baz){}
-
-    public record NumberDestRecord(Integer foo,
-                                   Long bar,
-                                   BigDecimal baz){}
 }
