+++ src/test/java/nablarch/core/beans/BeanUtilWithTypeParameterTest.java
@@ -44,8 +44,6 @@ public class BeanUtilWithTypeParameterTest {
         }
     }
 
-    public record ItemRecord<D extends Serializable>(List<D> items){}
-
     @Test
     public void testCreateAndCopyForBad() {
         Map<String, Object> map = Map.of(
@@ -71,20 +69,4 @@ public class BeanUtilWithTypeParameterTest {
         assertThat(form.getItems().get(0).getName(), is("aaa"));
         assertThat(form.getItems().get(1).getName(), is("bbb"));
     }
-
-
-    @Test
-    public void testCreateAndCopyForBadRecord() {
-        Map<String, Object> map = Map.of(
-                "items[0].name", "aaa",
-                "items[1].name", "bbb");
-
-        IllegalStateException result = assertThrows(IllegalStateException.class, () -> {
-            BeanUtil.createAndCopy(ItemRecord.class, map);
-        });
-        assertThat(result.getMessage(), is(
-                "BeanUtil does not support type parameter for List type, so the accessor in the concrete class must be overridden. "
-                        + "getter method = [nablarch.core.beans.BeanUtilWithTypeParameterTest$ItemRecord#items]"));
-    }
-
 }
